pyproject.toml
[project]
name = "dndrpg"
version = "0.1.0"
description = "D&D 3.5e text RPG (Effects/State engine + TUI)"
requires-python = ">=3.11"
dependencies = [
  "textual>=0.56",
  "rich>=13.7",
  "pydantic>=2.6",
  "pyyaml>=6.0",
  "typer>=0.12",
  "py-expression-eval>=0.3",
  "watchdog>=4.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "hypothesis>=6",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]